Add capability for relative file urls
Previously, the url crate's implementation of serde traits
required all the urls to be absolute urls. This commit provides an
alternative serde implementation for urls, which allow for urls to
be relative. The library exposes a static variable for a base url,
and yafpm-build fills in that static with the canonicalized path
to the config file.

Signed-off-by: John Arnold <code@iohannes.us>