Change build process env var logic to include deps
Previously, the build environment mounted the dependencies but the
only way to access them was with an absolute path. Now the path
to the dependency is given in an environment variable equal to the
dependency name. Also, additional env variables are now associated
with the Package struct, not the Build Context.

Signed-off-by: John Arnold <code@iohannes.us>